Modularized the code (part 2)
* Part 2 of modularizing the code
