+++ main.go
@@ -4,20 +4,21 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/Neo2308/indianrailways-gtfs/server"
 	"github.com/gorilla/mux"
 )
 
 func main() {
 
-	server := NewServer()
-	server.Setup()
+	irServer := server.NewServer()
+	irServer.Setup()
 	r := mux.NewRouter()
-	r.HandleFunc("/map", server.handleGetMap).Methods("GET")
-	r.HandleFunc("/map/{trainNumber}", server.handleGetMapForTrain).Methods("GET")
-	r.HandleFunc("/map/search/{prefixText}", server.handleGetMapBySearch).Methods("GET")
-	r.HandleFunc("/map/liveStation/{stationCode}", server.handleGetMapByLiveStation).Methods("GET")
-	r.HandleFunc("/map/show/all", server.populateAllTrains).Methods("GET")
-	r.HandleFunc("/save", server.saveGTFS).Methods("GET")
+	r.HandleFunc("/map", irServer.HandleGetMap).Methods("GET")
+	r.HandleFunc("/map/{trainNumber}", irServer.HandleGetMapForTrain).Methods("GET")
+	r.HandleFunc("/map/search/{prefixText}", irServer.HandleGetMapBySearch).Methods("GET")
+	r.HandleFunc("/map/liveStation/{stationCode}", irServer.HandleGetMapByLiveStation).Methods("GET")
+	r.HandleFunc("/map/show/all", irServer.PopulateAllTrains).Methods("GET")
+	r.HandleFunc("/save", irServer.SaveGTFS).Methods("GET")
 	err := http.ListenAndServe(":8080", r)
 	fmt.Println(err)
 }
