+++ server/convUtils.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 func getServiceId(runningDays string) string {
 	serviceId := ""
